修改编码器精度

+++ components/Drivers/AS5600.c
@@ -15,7 +15,7 @@
 #define ACK_VAL					0x0         /*!< I2C ack value */
 #define NACK_VAL				0x1         /*!< I2C nack value */
 
-static uint16_t AS5600Angle=0;
+static float AS5600Angle = 0.0f;
 
 //************************// 
 //  功能描述: AS5600_ IO 初始化函数
@@ -123,11 +123,11 @@ void AS5600_UpdateAngle()
 	memset(Angle.B08,0,sizeof(B16_B08));
 	AS5600_ReadData(RAW_ANGLE_L_REG,1,&Angle.B08[0]);
 	AS5600_ReadData(RAW_ANGLE_H_REG,1,&Angle.B08[1]);
-	AS5600Angle = Angle.B16*360/4096;
+	AS5600Angle = (float)Angle.B16*360.0f/4096.0f;
 }
 
 
-uint16_t AS5600_Angle()
+float AS5600_Angle()
 {
 	return AS5600Angle;
 }
@@ -145,7 +145,7 @@ void AS5600_Test()
 		// AS5600_ReadData(RAW_ANGLE_H_REG,1,&Angle.B08[1]);
 		// AngleTmp = Angle.B16*360/4096;
 		// AS5600Angle = AngleTmp;
-		printf("Angle:%d\r\n",AS5600_Angle());
+		printf("Angle:%.3f\r\n",AS5600_Angle());
 		vTaskDelayUntil(&Time,1/portTICK_PERIOD_MS);
     }
